Add files via upload
This version of procedural programming is a prototype of what a coffee-machine does in real life. A customer gets prompted, and selects a flavor of coffee, the coffee-machine processes the request and checks the resources availability, if resources are available and enough, the coffee is processed, if not, the customer, get a feedback.